SWC-6390: synapse-react-client - Refactor Login - Refactor Login to be a function component. Drop incomplete functionality. - Move authentication-related components to new folder
