+++ packages/synapse-react-client/demo/containers/App.tsx
@@ -2,7 +2,7 @@ import $RefParser from '@apidevtools/json-schema-ref-parser'
 import * as React from 'react'
 import { Alert } from 'react-bootstrap'
 import { BrowserRouter as Router, Link, Route } from 'react-router-dom'
-import Login from '../../src/lib/containers/Login'
+import Login from '../../src/lib/containers/auth/Login'
 import SynapseForm from '../../src/lib/containers/synapse_form_wrapper/SynapseForm'
 import SynapseFormSubmissionsGrid from '../../src/lib/containers/synapse_form_wrapper/SynapseFormSubmissionsGrid'
 import SynapseFormWrapper from '../../src/lib/containers/synapse_form_wrapper/SynapseFormWrapper'
