+++ packages/synapse-react-client/src/lib/umd.index.ts
@@ -23,8 +23,8 @@ import { HasAccessV2 as HasAccess } from './containers/access_requirements/HasAc
 import { HelpPopover } from './containers/HelpPopover'
 import ProjectViewCarousel from './containers/home_page/project_view_carousel/ProjectViewCarousel'
 import IconSvg from './containers/IconSvg'
-import Login from './containers/Login'
-import LoginPage from './containers/LoginPage'
+import Login from './containers/auth/Login'
+import LoginPage from './containers/auth/LoginPage'
 import PageProgress from './containers/PageProgress'
 import { AccessTokenPage } from './containers/personal_access_token/AccessTokenPage'
 import PlotlyWrapper from './containers/PlotlyWrapper'
